refactor(sci-clubs): move to hooks dir
