+++ lib/features/science_club/science_club_detail_view/hooks/use_detail_swipe.dart
